move field into to AbstractObjectSerializer
